+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v ?.??.?.? - ?? ??? ????
 - MD5Materials can be read from a PATH location
 - Expanded UBO support to more cleanly set uniforms within buffer block
+- teapot supports normal mapping
 
 ## v 5.25.0.0 - 09 Dec 2025
 - MD5Models support normal mapping